pyproject.toml
[tool.poetry]
name = "homework27"
version = "0.1.0"
description = ""
authors = ["AnastasiaGoryacheva <nastyaitaly@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
psycopg2 = "^2.9.3"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
